LPS-61952 Rename com.liferay.portal.service.configuration to com.liferay.portal.service.kernel.configuration
