+++ portal-service/src/com/liferay/portal/kernel/service/configuration/configurator/packageinfo
